split transcriptions tests
